Added scroll directive and changed the way scroll was performed.
Instead of using document.getElementById in the Typescript code to find
the element and scroll to it, the appScroll directive is included in the
HTML div element that renders when the element is loading.
This way, when that element is rendered, the appScroll directive is executed,
and the directive scrolls the page to that HTML element.