Refactored text options to a component.
This component will be refactored later into 3 components.

The TextSectionComponent doesn't need now to inject the child components
using @ViewChild, because it isn't going to use them directly in the Typescript.
The TextOptionsComponent is the one that is going to use the rest of the
sibling components, so they are injected into the TextOptionsComponent using
@Input and passing them in the HTML.