Refactored topics in text format to a component.
There is a bug: when the topics configuration updates the number
of keywords in text format for the first time, the table doesn't
refresh, because the @Input property that holds the number of keywords
isn't updated before the call to the service.
This will be fixed when the topics-options component is created.