Modified h1 and h2 to make them adaptable to extra small screen sizes.
The h1 in extra small screens uses vw, sets the background-color to blue
and reduces the top margin to 0.

The h2 in extra small screens uses vw and % margin.

vw is relative to 1% of the width of the viewport.

Also sliders min and max values doesn't have margin in extra small screens.